+++ tests/Unit/ProcessTest.php
@@ -12,7 +12,7 @@ use Queue\Driver\Connection;
 use QueueTest\Fake\ConnectionFake;
 use QueueTest\Fake\ConsumerFake;
 use QueueTest\Fake\ConsumerFakeWithRetry;
-use QueueTest\Fake\ProducerFake;
+use QueueTest\Mocks\Producer\ProducerMock;
 use QueueTest\Mocks\Consumer\ConsumerFakeMessage;
 use QueueTest\Mocks\Consumer\ConsumerFakeWithNoAck;
 
@@ -25,7 +25,7 @@ class ProcessTest extends \PHPUnit_Framework_TestCase
     public function testProducerAndConsumerFlux(Connection $connection)
     {
         $messageTxt = 'testMessage';
-        $producer = new ProducerFake($connection);
+        $producer = new ProducerMock($connection);
         $message = $producer->prepare($messageTxt);
         $producer->publish($message);
 
@@ -41,7 +41,7 @@ class ProcessTest extends \PHPUnit_Framework_TestCase
     public function testConsumerNoAckFlux(Connection $connection)
     {
         $messageTxt = 'testMessageWithNoAck';
-        $producer = new ProducerFake($connection);
+        $producer = new ProducerMock($connection);
         $message = $producer->prepare($messageTxt);
         $producer->publish($message);
 
@@ -57,7 +57,7 @@ class ProcessTest extends \PHPUnit_Framework_TestCase
      */
     public function testConsumerRetryFlux(Connection $connection)
     {
-        $producer = new ProducerFake($connection);
+        $producer = new ProducerMock($connection);
         $message = $producer->prepare('');
         $producer->publish($message);
 
@@ -71,7 +71,7 @@ class ProcessTest extends \PHPUnit_Framework_TestCase
      */
     public function testMessageAccessor(Connection $connection)
     {
-        $producer = new ProducerFake($connection);
+        $producer = new ProducerMock($connection);
         $message = $producer->prepare('');
         $producer->publish($message);
